refactor(frontend): rename Ethereum send amount component
