+++ src/frontend/src/eth/components/send/EthSendAmount.svelte
