+++ src/frontend/src/eth/components/send/SendForm.svelte
@@ -2,7 +2,7 @@
 	import { isNullish } from '@dfinity/utils';
 	import { createEventDispatcher, getContext } from 'svelte';
 	import FeeDisplay from '$eth/components/fee/FeeDisplay.svelte';
-	import SendAmount from '$eth/components/send/SendAmount.svelte';
+	import EthSendAmount from '$eth/components/send/EthSendAmount.svelte';
 	import SendDestination from '$eth/components/send/SendDestination.svelte';
 	import SendInfo from '$eth/components/send/SendInfo.svelte';
 	import SendNetworkICP from '$eth/components/send/SendNetworkICP.svelte';
@@ -53,7 +53,7 @@
 			<SendNetworkICP {destination} {sourceNetwork} bind:network />
 		{/if}
 
-		<SendAmount {nativeEthereumToken} bind:amount bind:insufficientFunds />
+		<EthSendAmount {nativeEthereumToken} bind:amount bind:insufficientFunds />
 
 		<SendSource
 			token={$sendToken}
